third_party/blink/web_tests/http/tests/security/contentTypeOptions/nosniff-dynamic-script-blocked-expected.txt

CONSOLE ERROR: Refused to execute script from 'http://127.0.0.1:8000/security/contentTypeOptions/resources/script-with-header.pl?mime=application/json' because its MIME type ('application/json') is not executable, and strict MIME type checking is enabled.
Check that script sent with an 'X-Content-Type-Options: nosniff' header is correctly blocked if the MIME type isn't scripty.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.scriptsSuccessfullyLoaded is 0
PASS successfullyParsed is true

TEST COMPLETE
